requirements.txt
streamlit
pandas
pillow
pytz

